🎨 Style : CvCustomInput
